test: complete Step 2.2 - create workflow core tests
- Created test_workflow_models.py with 4 test classes
- TestRunWorkflowModels, TestRunSpecificWorkflowModels, TestGetWorkflowRunDetailModels, TestStopWorkflowModels
- All tests pass with minimal essential code
- Backed up original file as test_workflow_models_original.py
- Updated progress: Phase 2 now 2/14 steps (14.3%)
