+++ docs/tests/tests-todo.md
@@ -41,7 +41,7 @@
 
 ### Analysis and Core Implementation
 - [x] **Step 2.1**: Analyze Current Workflow Tests
-- [ ] **Step 2.2**: Create Workflow Core Tests
+- [x] **Step 2.2**: Create Workflow Core Tests
 - [ ] **Step 2.3**: Test Workflow Core Implementation
 - [ ] **Step 2.4**: Create Workflow Public Models Tests
 - [ ] **Step 2.5**: Test Workflow Public Models
@@ -99,12 +99,12 @@
 
 ### Completion Status
 - **Phase 1**: 18/18 steps completed (100.0%)
-- **Phase 2**: 1/14 steps completed (7.1%)
+- **Phase 2**: 2/14 steps completed (14.3%)
 - **Phase 3**: 0/12 steps completed (0%)
 - **Phase 4**: 0/3 steps completed (0%)
 
 ### Overall Progress
-- **Total Steps**: 19/47 completed (40.4%)
+- **Total Steps**: 20/47 completed (42.6%)
 - **Files to Refactor**: 3 large test files
 - **Expected New Files**: ~30+ test files
 - **Estimated Line Reduction**: ~60-70%
